+++ tests/__init__.py
@@ -28,4 +28,6 @@ class BaseWebTest(CoreWebTest):
         config = configparser.ConfigParser()
         config.read(ini_path)
         settings = dict(config.items('app:main'))
+        if extras:
+            settings.update(extras)
         return settings