+++ tests/test_elasticsearch.py
@@ -40,6 +40,50 @@ class PluginSetup(BaseWebTest, unittest.TestCase):
             assert not resp.json["elasticsearch"]
 
 
+class PostActivation(BaseWebTest, unittest.TestCase):
+
+    @classmethod
+    def get_app_settings(cls, extras=None):
+        settings = super().get_app_settings(extras)
+        settings['storage_backend'] = 'kinto.core.storage.postgresql'
+        settings['storage_url'] = 'postgres://postgres:postgres@localhost:5432/postgres'
+        settings['permission_backend'] = 'kinto.core.permission.postgresql'
+        settings['permission_url'] = settings['storage_url']
+        return settings
+
+    def setUp(self):
+        app = self.make_app(settings={"kinto.includes": ""})
+        capabilities = app.get("/").json["capabilities"]
+        assert "elasticsearch" not in capabilities
+
+        app.put("/buckets/bid", headers=self.headers)
+        app.put("/buckets/bid/collections/cid", headers=self.headers)
+        app.post_json("/buckets/bid/collections/cid/records",
+                      {"data": {"before": "indexing"}},
+                      headers=self.headers)
+
+    def test_search_does_not_fail(self):
+        resp = self.app.get("/buckets/bid/collections/cid/records",
+                            headers=self.headers)
+        assert len(resp.json["data"]) == 1
+
+        resp = self.app.get("/buckets/bid/collections/cid/search",
+                            headers=self.headers)
+        results = resp.json
+        assert len(results["hits"]["hits"]) == 0
+
+    def test_record_creation_does_not_fail(self):
+        self.app.post_json("/buckets/bid/collections/cid/records",
+                           {"data": {"after": "indexing"}},
+                           headers=self.headers)
+
+        resp = self.app.get("/buckets/bid/collections/cid/search",
+                            headers=self.headers)
+        results = resp.json
+        assert len(results["hits"]["hits"]) == 1
+        assert results["hits"]["hits"][0]["_source"]["after"] == "indexing"
+
+
 class RecordIndexing(BaseWebTest, unittest.TestCase):
 
     def setUp(self):