+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ Changelog
 
 - Only index records if the storage transaction is committed (fixes #15)
 - Do not allow to search if no read permission on collection or bucket (fixes #7)
+- Fix empty results response when plugin was enabled after collection creation (ref #20)
 
 **Internal changes**